Added Item Looking
You can know look at items by typing "look" (or "l") and then the item name. It will search for the item in both the player's and current location's items list. If nothing is found then the player is informed of that.
You are also now able to type "clear" or "c" to clear the screen and "exit" can also be used to quit the game.